requirements.txt
asgiref==3.6.0
atomicwrites==1.4.1
attrs==22.2.0
base64image==0.5.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.12
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==40.0.1
defusedxml==0.7.1
diff-match-patch==20200713
drf-extra-fields==3.2.1
Django==3.2
django-cors-headers==3.14.0
django-extra-fields==3.0.2
django-filter==22.1
django-import-export==3.0.2
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
et-xmlfile==1.1.0
idna==3.4
importlib-metadata==1.7.0
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.1.2
MarkupPy==1.14
MarkupSafe==2.1.2
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.0
Pillow==9.4.0
pluggy==0.13.1
py==1.11.0
pycparser==2.21
PyJWT==2.1.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-decouple==3.5
python-dotenv==0.21.0
python3-openid==3.2.0
psycopg2-binary==2.9.3
pytz==2022.7.1
PyYAML==6.0
requests==2.26.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.4.0
sqlparse==0.4.3
tablib==3.4.0
toml==0.10.2
typing-extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.15
xlrd==2.0.1
xlwt==1.3.0
zipp==3.15.0